+++ src/Listener/MaintenanceListener.php
@@ -1,0 +1,59 @@
+<?php
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\ConfigBundle\Listener;
+
+use c975L\ConfigBundle\Service\ConfigServiceInterface;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
+use Twig\Environment;
+use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
+
+#[AsEventListener(event: 'kernel.request', method: 'onKernelRequest', priority: 10)]
+class MaintenanceListener
+{
+    public function __construct(
+        private readonly ConfigServiceInterface $configService,
+        private readonly Environment $twig
+    ) {
+    }
+
+    public function onKernelRequest(RequestEvent $event): void
+    {
+        // Loads configs
+        $this->configService->loadAll();
+
+        // Maintenance mode
+        if (!$event->isMainRequest() || false === $this->configService->get("site-maintenance")) {
+             return;
+        }
+
+        // Access via token in URL : ?t=secret_token
+        $request = $event->getRequest();
+        if ($request->query->get('t') === $this->configService->get("site-maintenance-hash")) {
+            $maintenance = [
+                'access' => true,
+                'access_time' => time() + 3 * 60 * 60,
+            ];
+            $request->getSession()->set('site-maintenance', $maintenance);
+
+            return;
+        }
+
+        // Access via session (valid token)
+        $maintenance = $request->getSession()->get('site-maintenance');
+        if ($maintenance && $maintenance['access'] && $maintenance['access_time'] > time()) {
+            return;
+        }
+
+        // Otherwise maintenance page
+        $html = $this->twig->render('@c975Config/maintenance/index.html.twig');
+        $event->setResponse(new Response($html, 503));
+    }
+}