+++ src/Management/MenuProviderInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace c975L\ConfigBundle\Management;
+
+interface MenuProviderInterface
+{
+    public function getSection(): array;
+
+    public function getMenu(): array;
+
+    public function getMenuItems(): iterable;
+}